Refactor : 오타 + 변수명 및 경로 수정
