+++ src/pages/Admin/Artists.vue
