+++ src/router/index.ts
@@ -3,19 +3,19 @@ import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router';
 const routes: Array<RouteRecordRaw> = [
   {
     path: '/',
-    component: () => import('../pages/Home.vue'),
+    component: () => import('@/pages/Home.vue'),
   },
   {
     path: '/admin',
-    component: () => import('../pages/Admin.vue'),
+    component: () => import('@/pages/Admin/Admin.vue'),
   },
   {
     path: '/admin/group',
-    component: () => import('../pages/Group.vue'),
+    component: () => import('@/pages/Admin/Group.vue'),
   },
   {
     path: '/admin/artists',
-    component: () => import('../pages/Artists.vue'),
+    component: () => import('@/pages/Admin/Artists.vue'),
   },
 ];
